Return more informational errors to the client
Current we are outputting errors as text/plain (sometimes in the case of
application/json) and they are not very informative. It's a lot easier
for debugging purposes to be able to get more insight into the specific
errors, I don't think there's a big issue with information disclosure
here.

Instead of plaintext errors, I added a "message" field to the JSON
output also.